refactor from js to ts for app, config, server
