+++ src/common/config.ts
@@ -1,0 +1,17 @@
+import dotenv from 'dotenv';
+import path from 'path';
+
+dotenv.config({
+  path: path.join(__dirname, '../../.env')
+});
+
+const { PORT, NODE_ENV, MONGO_CONNECTION_STRING, JWT_SECRET_KEY, AUTH_MODE: AUTH_MODE_PROPS  } = process.env;
+const AUTH_MODE:boolean = AUTH_MODE_PROPS === 'true';
+
+export {
+  PORT,
+  NODE_ENV,
+  MONGO_CONNECTION_STRING,
+  JWT_SECRET_KEY,
+  AUTH_MODE,
+};
